resolve requested changes by @mrjameshamilton
