+++ base/src/main/java/proguard/analysis/CallResolver.java
@@ -51,8 +51,8 @@ import proguard.classfile.util.ClassUtil;
 import proguard.classfile.visitor.ClassVisitor;
 import proguard.classfile.visitor.LineNumberFinder;
 import proguard.evaluation.BasicInvocationUnit;
-import proguard.evaluation.EmptyCodeAttributeException;
-import proguard.evaluation.ExcessiveComplexityException;
+import proguard.evaluation.exception.EmptyCodeAttributeException;
+import proguard.evaluation.exception.ExcessiveComplexityException;
 import proguard.evaluation.ExecutingInvocationUnit;
 import proguard.evaluation.InvocationUnit;
 import proguard.evaluation.PartialEvaluator;
