+++ base/src/main/java/proguard/evaluation/exception/EmptyCodeAttributeException.java
@@ -15,8 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package proguard.evaluation;
 
+package proguard.evaluation.exception;
+
+import proguard.exception.ErrorId;
 import proguard.exception.ProguardCoreException;
 
 public class EmptyCodeAttributeException
@@ -25,7 +27,6 @@ public class EmptyCodeAttributeException
 
     public EmptyCodeAttributeException(String message)
     {
-        //ToDo: update component error id
-        super(3, message);
+        super(ErrorId.EMPTY_CODE_ATTRIBUTE, message);
     }
 }
